fix Yaspin._freeze to allow it to consume raw bytes
